drawing_window_python/requirements.txt
ai-edge-litert==2.1.0
backports.strenum==1.3.1
flatbuffers==25.12.19
numpy==2.2.6
protobuf==6.33.2
pycairo==1.29.0
PyGObject==3.50.0
tqdm==4.67.1
typing_extensions==4.15.0
